# docs/notebook/openapi.yaml
...
openapi: 3.0.3
info:
  title: Kyuubi Notebook API
  version: "1.0.0"
  description: >
    Notebook subsystem of Apache Kyuubi. Covers the endpoints implemented in PR 1: folders,
    notebooks, cells, revisions, permissions, schedules, import, export, search and the service
    endpoints. Runtime, execution and Python-environment endpoints are specified in API.md and
    are added to this document when they are implemented.

    Authentication reuses the Kyuubi REST frontend; under kyuubi.authentication=OIDC that is an
    OAuth2 bearer token. There is no notebook-specific login, and no request body carries owner
    or audit fields - they are derived from the authenticated caller.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /notebook-folders:
    post:
      tags: [Notebook]
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "200":
          description: The created folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotebookFolder"
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      tags: [Notebook]
      summary: List the caller's folders
      parameters:
        - name: parentId
          in: query
          required: false
          schema: { type: string }
          description: Empty string lists roots; omitted lists all folders.
      responses:
        "200":
          description: Folders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NotebookFolder" }
  /notebook-folders/{folderId}:
    parameters:
      - name: folderId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Notebook]
      summary: Get a folder
      responses:
        "200":
          description: The folder
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotebookFolder" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [Notebook]
      summary: Rename or reparent a folder, rewriting descendant paths
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateFolderRequest" }
      responses:
        "200":
          description: The updated folder
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotebookFolder" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [Notebook]
      summary: Soft-delete a folder and its subtree
      parameters:
        - name: version
          in: query
          required: false
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Error" }
  /notebooks:
    post:
      tags: [Notebook]
      summary: Create a notebook
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateNotebookRequest" }
      responses:
        "200":
          description: The created notebook
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notebook" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      tags: [Notebook]
      summary: List notebooks the caller may read
      parameters:
        - { name: owner, in: query, schema: { type: string } }
        - { name: folderId, in: query, schema: { type: string } }
        - { name: name, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200":
          description: A page of notebooks
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotebookPage" }
  /notebooks/{notebookId}:
    parameters:
      - name: notebookId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Notebook]
      summary: Get a notebook
      parameters:
        - name: includeCells
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: The notebook
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notebook" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [Notebook]
      summary: Update notebook attributes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateNotebookRequest" }
      responses:
        "200":
          description: The updated notebook
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notebook" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [Notebook]
      summary: Soft-delete a notebook, dropping its grants and schedule
      parameters:
        - name: version
          in: query
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Error" }
...
  /notebooks:import:
    post:
      tags: [Notebook]
      summary: Import a document; ids are regenerated and the caller becomes the owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                format: { type: string, enum: [KYUUBI, IPYNB] }
                name: { type: string }
                folderId: { type: string, nullable: true }
                content: { type: string, description: The document as raw JSON text }
      responses:
        "200":
          description: The imported notebook
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notebook" }
        "400": { $ref: "#/components/responses/Error" }
  /notebooks:search:
    get:
      tags: [Notebook]
      summary: Search names, descriptions and cell sources within authorized notebooks
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: folderId, in: query, schema: { type: string } }
        - name: language
          in: query
          schema: { type: string, enum: [SQL, PYTHON, MARKDOWN] }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200":
          description: A page of notebooks
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotebookPage" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: An error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable machine-readable code; see API.md for the full list.
            message:
              type: string
              description: Safe for display; never contains a stack trace, secret or path.
            requestId: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: { type: string } }
    NotebookFolder:
      type: object
      properties:
        id: { type: string }
        parentId: { type: string, nullable: true }
        name: { type: string }
        path:
          type: string
          description: Derived as /<owner>/<ancestors>/<name>; never client-supplied.
        owner: { type: string }
        createdAt: { type: integer, format: int64 }
        createdBy: { type: string }
        updatedAt: { type: integer, format: int64 }
        updatedBy: { type: string }
        version: { type: integer, format: int64 }
    Notebook:
      type: object
      properties:
        id: { type: string }
        folderId: { type: string, nullable: true }
        path: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        owner: { type: string }
        defaultCatalog: { type: string, nullable: true }
        defaultSchema: { type: string, nullable: true }
        runtimeProfile: { type: string, nullable: true }
        formatVersion: { type: integer }
        createdAt: { type: integer, format: int64 }
        createdBy: { type: string }
        updatedAt: { type: integer, format: int64 }
        updatedBy: { type: string }
        version: { type: integer, format: int64 }
        role:
          type: string
          enum: [OWNER, EDITOR, VIEWER]
          nullable: true
          description: The calling principal's effective role.
        cells:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/NotebookCell" }
    NotebookCell:
      type: object
      properties:
        id: { type: string }
        notebookId: { type: string }
        position: { type: integer }
        cellType: { type: string, enum: [CODE, MARKDOWN] }
        language: { type: string, enum: [SQL, PYTHON, MARKDOWN] }
        source: { type: string }
        metadata: { type: object, additionalProperties: { type: string } }
        configuration: { type: object, additionalProperties: { type: string } }
        createdAt: { type: integer, format: int64 }
        updatedAt: { type: integer, format: int64 }
        version: { type: integer, format: int64 }
...
    NotebookPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Notebook" }
        nextCursor:
          type: string
          nullable: true
          description: Opaque; pass back verbatim. Do not parse.
        hasMore: { type: boolean }
    CreateFolderRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        parentId: { type: string, nullable: true }
    UpdateFolderRequest:
      type: object
      properties:
        name: { type: string }
        parentId:
          type: string
          nullable: true
          description: Empty string moves the folder to the caller's root.
        version: { type: integer, format: int64 }
    CreateNotebookRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        folderId: { type: string, nullable: true }
        description: { type: string }
        defaultCatalog: { type: string }
        defaultSchema: { type: string }
        runtimeProfile: { type: string }
        cells:
          type: array
          items: { $ref: "#/components/schemas/CreateCellRequest" }
    UpdateNotebookRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        defaultCatalog: { type: string }
        defaultSchema: { type: string }
        runtimeProfile: { type: string }
        version: { type: integer, format: int64 }
    CreateCellRequest:
      type: object
      properties:
        cellType: { type: string, enum: [CODE, MARKDOWN], default: CODE }
        language: { type: string, enum: [SQL, PYTHON, MARKDOWN] }
        source: { type: string }
        position: { type: integer, description: Defaults to the end of the notebook. }
        metadata: { type: object, additionalProperties: { type: string } }
        configuration: { type: object, additionalProperties: { type: string } }
...